Restructure huge auth file and user file
This breaks up the auth action file
It also breaks up the user controller
