+++ app/actions/modals.js
@@ -61,3 +61,33 @@ export function changeWalkThroughState(walkThroughState) {
         state: walkThroughState,
     }
 }
+
+export function  walkThroughFinished(email, token) {
+    return (dispatch) => {
+        return fetch('/account/walkthrough', {
+            method: 'put',
+            headers: {
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}`
+            },
+            body: JSON.stringify({
+                email: email,
+            })
+        }).then((response) => {
+            if (response.ok) {
+                return response.json().then((json) => {
+                    dispatch({
+                        type: 'WALK_THROUGH_FINISHED',
+                        user: json.user
+                    });
+                });
+            } else {
+                return response.json().then((json) => {
+                    dispatch({
+                        type: 'WALK_THROUGH_FINISH_FAILURE'
+                    });
+                });
+            }
+        });
+    };
+}
